+++ changelog.md
@@ -4,3 +4,8 @@
 
 * Added Data.Binary interface for GHC 7.6
 
+## 0.0.5.0
+
+* changed the Functor instance of Data.RTree.Strict to be strict
+
+* Data.RTree.Strict.RTree is now a newtype of Data.RTree.RTree